Revert "chore: 🤖 adds vite-note as dep"
