+++ package.json
@@ -5,7 +5,7 @@
   "main": "src/index.ts",
   "author": "Askrella Software Agency UG (haftungsbeschränkt)",
   "scripts": {
-    "start": "vite-node --watch src/index.ts"
+    "start": "npx vite-node src/index.ts",
     "format": "prettier --write \"src/**/*.*\""
   },
   "license": "MIT",
@@ -15,7 +15,6 @@
     "openai": "^3.1.0",
     "prettier": "^2.8.4",
     "qrcode-terminal": "^0.12.0",
-    "vite-node": "^0.28.5",
     "whatsapp-web.js": "^1.19.4"
   }
 }
