+++ src/dalle.ts
@@ -1,7 +1,6 @@
 import process from "process";
 import { Configuration, OpenAIApi } from "openai";
-
-import { MessageMedia } from "whatsapp-web.js";
+const { MessageMedia } = require("whatsapp-web.js");
 
 // Environment variables
 require("dotenv").config();
