+++ src/gpt.ts
@@ -1,6 +1,5 @@
-import process from "process";
+const process = require("process");
 import { ChatGPTAPI, ChatMessage } from "chatgpt";
-import { type Message } from "whatsapp-web.js";
 
 // Environment variables
 require("dotenv").config();
@@ -13,7 +12,7 @@ const api = new ChatGPTAPI({
 // Mapping from number to last conversation id
 const conversations = {};
 
-export const handleMessageGPT = async (message: Message, prompt: string) => {
+const handleMessageGPT = async (message: any, prompt: any) => {
   try {
     // Get last conversation
     const lastConversation = conversations[message.from];
@@ -26,7 +25,6 @@ export const handleMessageGPT = async (message: Message, prompt: string) => {
 
     // Check if we have a conversation with the user
     let response: ChatMessage;
-
     if (lastConversation) {
       // Handle message with previous conversation
       response = await api.sendMessage(prompt, lastConversation);
@@ -58,3 +56,5 @@ export const handleMessageGPT = async (message: Message, prompt: string) => {
     );
   }
 };
+
+export { handleMessageGPT };
