+++ src/index.ts
@@ -1,42 +1,67 @@
-// Whatsapp client
-import { Client, LocalAuth } from "whatsapp-web.js";
+const process = require("process");
+const qrcode = require("qrcode-terminal");
+const { Client } = require("whatsapp-web.js");
 
-import {
-  onMessage,
-  onReady,
-  onQRCode,
-  onAuthenticated,
-  onMessageCreate,
-} from "./listeners";
+// ChatGPT & DALLE
+import { handleMessageGPT } from "./gpt";
+import { handleMessageDALLE } from "./dalle";
 
 // Environment variables
 require("dotenv").config();
+
+// Prefixes
+const prefixEnabled = process.env.PREFIX_ENABLED == "true";
+const gptPrefix = "!gpt";
+const dallePrefix = "!dalle";
 
 // Whatsapp Client
 const client = new Client({
   puppeteer: {
     args: ["--no-sandbox"],
   },
-  authStrategy: new LocalAuth(),
 });
 
 // Entrypoint
 const start = async () => {
-  client.on("qr", onQRCode);
-  client.on("authenticated", onAuthenticated);
-  client.on("ready", onReady);
+  // Whatsapp auth
+  client.on("qr", (qr: string) => {
+    console.log("[Whatsapp ChatGPT] Scan this QR code in whatsapp to log in:");
+    qrcode.generate(qr, { small: true });
+  });
 
-  client.on("message", onMessage);
-  client.on("message_create", onMessageCreate);
+  // Whatsapp ready
+  client.on("ready", () => {
+    console.log("[Whatsapp ChatGPT] Client is ready!");
+  });
+
+  // Whatsapp message
+  client.on("message", async (message: any) => {
+    const messageString = message.body;
+    if (messageString.length == 0) return;
+    if (message.from == "status@broadcast") return;
+
+    if (prefixEnabled) {
+      // GPT (!gpt <prompt>)
+      if (messageString.startsWith(gptPrefix)) {
+        const prompt = messageString.substring(gptPrefix.length + 1);
+        await handleMessageGPT(message, prompt);
+        return;
+      }
+
+      // DALLE (!dalle <prompt>)
+      if (messageString.startsWith(dallePrefix)) {
+        const prompt = messageString.substring(dallePrefix.length + 1);
+        await handleMessageDALLE(message, prompt);
+        return;
+      }
+    } else {
+      // GPT (only <prompt>)
+      await handleMessageGPT(message, messageString);
+    }
+  });
 
   // Whatsapp initialization
-  return client.initialize();
+  client.initialize();
 };
 
-start()
-  .then(() => {
-    console.log("[Whatsapp ChatGPT] Running");
-  })
-  .catch((error: any) => {
-    console.error("An error happened:", error);
-  });
+start();
