--- tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "compilerOptions": {
-    "module": "commonjs",
-    "target": "es5",
-    "esModuleInterop": true
-  },
-  "exclude": ["node_modules"]
-}
